[REFACTOR] star-log 구조 분리
